requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.0
fonttools==4.55.0
fsspec==2025.5.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
mariadb==1.1.11
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.1.3
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.16.0
sympy==1.14.0
tifffile==2025.6.1
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.14.0
tzdata==2024.2
Werkzeug==3.1.3
